[Metrics] Histogram for deadline remaining on Append (#737)
In https://github.com/sigstore/rekor-tiles/issues/434 it is reported that there are deadline exceeded errors. This could be Tessera taking longer than expected, or it could be clients calling this with unreasonably short deadlines. This histogram metric should give some insight into the spectrum of deadlines that clients are expecting Tessera to work with.