fix missing env var in example in readme (#785)
* fix missing env var in example in readme

Set the "LOG_DIR" var (already used in the example script) to a simple dummy value which is likely to work, so that the example script as a whole now fully runs with no modification needed.

* further simplify example

Co-locate all variables needing changing at the top, for consistent and easy visibility.

Elide a mkdir because the tooling will do it automatically.