Use local stored checkpoint as witnessing size hint (#796)
Changes the witnessing code to use the local published checkpoint as a hint for the size the witnesses are likely to know about.

Previously, the WitnessGateway used the size of the last checkpoint it was provided with, however this heuristic fails when there is more than one instance of an application frontend, and the cost is that we end up doing up to 2x as many requests to witnesses as strictly necessary.