Allow control over publication of unchanged checkpoints (#807)
Adds an option (and support in the POSIX backend) for controlling how frequently (if at all) checkpoints should be re-published when the log size hasn't grown.

This is helpful in reducing load on witnesses in the case where the log is quiescent, as well as any other application which requires further action to take place whenever a new checkpoint is published.